some edits in the code:
Look at the commit:
(main point is #3)

1) I am not sure what is ceil() in C. If you do x / y where x and y are int,
then the results is an int, and this is int is ``floor(x/y)`` in matlab. So I
changed your ceil(x/y) by x/y.  You need to understand this. ``x/y`` in C takes
with two (int)'s returns an (int) and this is (int) is ``floor(x/y)``.

2) I went all the way down to n = 1 in the recursion.
What I would like you to do now, is to have two leaves.
One leaf is for n < 10 say, and then you call an unblocked code like
	info = lila_dgeqr2_w02b( m, n, i, mt, A, lda, T, ldt, Q, ldq, work, lwork );
fine. The leaf for n < 2 (so n = 1), needs to be done by hand.  ``All`` you
need to do is construct V, Q and tau. (You have two tau: one in T and one in
TTT)

3) I am using two recursions, one on the left, one on the right, you were only
using recursion on the right. I am not sure why.