adding the code lila_dgeqr1_w03a and lila_dgeqr1_w03a
These codes only work for n=1. (This why the name is DGEQR1 with a 1 for n=1.)

W03a relies on LAPACK subroutines, W03b is done ``by hand``.

For now we have T and TTT in the interface. The goal is to only have TTT once we have BZ working